Add better error handling for catchup subscriptions
When event handler crashes, we should call error handler and silently hang on subscription to wait until the handler is fixed
before processing next events. We don not want to have inconsistent state, but we do not want to crash listening method. When handler is fixed, we can
continue from the failing point